refactor: NodeResult.from_context/from_envelope; move type to models/
Replace the `deserialize_to_node_result(envelope, ...)` free function with
alternative constructors on the type:
- `NodeResult.from_context(ctx, ...)` — core projector, sources identity from a
  stamped SessionRunContext (correlation_id falls back to ctx.correlation_id).
- `NodeResult.from_envelope(envelope, ...)` — thin delegate over from_context.

Move the client-facing NodeResult plus the `_UNSET` sentinel and `_extract_*`
helpers from calfkit/client/{node_result,deserialize}.py into
calfkit/models/node_result.py, so the client and consumer nodes both depend
"down" on the shared models kernel instead of nodes importing "up" from client.

Not re-exported from calfkit.models (the package already exports a distinct
control-flow action-union NodeResult); importers use the full submodule path.
Public surfaces `calfkit.NodeResult` and `calfkit.client.NodeResult` are
preserved via re-export. Pure rename/move/dedup — no behavior change; adds
from_context/from_envelope/delegation tests.
