+++ calfkit/client/invocation_handle.py
@@ -5,9 +5,8 @@ from dataclasses import dataclass, field
 from typing import Any, Generic
 
 from calfkit._types import OutputT
-from calfkit.client.deserialize import _UNSET, deserialize_to_node_result
-from calfkit.client.node_result import NodeResult
 from calfkit.models.envelope import Envelope
+from calfkit.models.node_result import _UNSET, NodeResult
 
 
 @dataclass
@@ -53,4 +52,4 @@ class InvocationHandle(Generic[OutputT]):
             envelope = await asyncio.wait_for(self._future, timeout=timeout)
         else:
             envelope = await self._future
-        return deserialize_to_node_result(envelope, self._output_type, correlation_id=self.correlation_id)
+        return NodeResult.from_envelope(envelope, self._output_type, correlation_id=self.correlation_id)
