+++ tests/test_lifecycle_review_fixes.py
@@ -125,7 +125,7 @@ def test_tool_context_default_resources_is_a_mapping() -> None:
 def test_node_result_default_resources_is_a_mapping() -> None:
     from collections.abc import Mapping
 
-    from calfkit.client.node_result import NodeResult
+    from calfkit.models.node_result import NodeResult
     from calfkit.models.state import State
 
     result = NodeResult(output=None, state=State(), correlation_id="cid")
